Revert "enable building on intel mac"
This reverts commit 3ec40fc30ada41364298e70267366d2aa6dbff43.

This commit has an error, it uses ${ARCHS_STANDARD} instead of the correct
value of $(ARCHS_STANDARD). The reason it still worked was ${ARCHS_STANDARD}
is a CMake variable, and is set to "" by default.

The parentheses version will work correctly if using Xcode or Makefiles as your
build system, but fails with ninja (default build system for CLion).